// apps/web/src/constants/typechain/IManager.ts
export declare namespace IManager {
  export type FounderParamsStruct = {
    wallet: PromiseOrValue<string>;
    ownershipPct: PromiseOrValue<BigNumberish>;
    vestExpiry: PromiseOrValue<BigNumberish>;
  };

  export type FounderParamsStructOutput = [string, BigNumber, BigNumber] & {
    wallet: string;
    ownershipPct: BigNumber;
    vestExpiry: BigNumber;
  };

  export type TokenParamsStruct = { initStrings: PromiseOrValue<BytesLike> };

  export type TokenParamsStructOutput = [string] & { initStrings: string };

  export type AuctionParamsStruct = {
    reservePrice: PromiseOrValue<BigNumberish>;
    duration: PromiseOrValue<BigNumberish>;
  };

  export type AuctionParamsStructOutput = [BigNumber, BigNumber] & {
    reservePrice: BigNumber;
    duration: BigNumber;
  };

  export type GovParamsStruct = {
    timelockDelay: PromiseOrValue<BigNumberish>;
    votingDelay: PromiseOrValue<BigNumberish>;
    votingPeriod: PromiseOrValue<BigNumberish>;
    proposalThresholdBps: PromiseOrValue<BigNumberish>;
    quorumThresholdBps: PromiseOrValue<BigNumberish>;
    vetoer: PromiseOrValue<string>;
  };

  export type GovParamsStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    string
  ] & {
    timelockDelay: BigNumber;
    votingDelay: BigNumber;
    votingPeriod: BigNumber;
    proposalThresholdBps: BigNumber;
    quorumThresholdBps: BigNumber;
    vetoer: string;
  };
}